# specs/001-tharwa-platform-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tharwa API
  version: 1.0.0
  description: >
    Backend for Tharwa (ثروة). Public read endpoints are unauthenticated for MVP market browsing.
    Admin routes are under /admin/v1 (separate security scheme). Parallel EGP is not in MVP.
servers:
  - url: https://api.example.com
    description: Production (replace with real host)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/watchlist:
    get:
      tags: [Consumer]
      summary: List watchlist
      operationId: listWatchlist
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Consumer]
      summary: Add instrument to watchlist
      operationId: addWatchlistItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistAddRequest"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                type: object
                required: [item]
                properties:
                  item:
                    $ref: "#/components/schemas/WatchlistItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/watchlist/items/{id}:
    delete:
      tags: [Consumer]
      summary: Remove watchlist item
      operationId: deleteWatchlistItem
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/watchlist/reorder:
    patch:
      tags: [Consumer]
      summary: Reorder watchlist
      operationId: reorderWatchlist
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistReorderRequest"
      responses:
        "200":
          description: Full ordered list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/zakat/nisab:
    get:
      tags: [Market]
      summary: Current nisab threshold from 21k gold gram price
      operationId: getZakatNisab
      responses:
        "200":
          description: Nisab snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZakatNisab"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/zakat/compute:
    post:
      tags: [Market]
      summary: Compute estimated zakat from asset inputs
      operationId: computeZakat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZakatComputeRequest"
      responses:
        "200":
          description: Zakat estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZakatComputeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/zakat/prefill:
    get:
      tags: [Consumer]
      summary: Prefill equity values from trade journal portfolio
      operationId: getZakatPrefill
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Journal-based prefill lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZakatPrefill"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/zakat/sessions:
    get:
      tags: [Consumer]
      summary: List saved zakat calculation sessions
      operationId: listZakatSessions
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Recent sessions (newest first)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ZakatSessionSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Consumer]
      summary: Save a zakat calculation session
      operationId: createZakatSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZakatSessionCreate"
      responses:
        "201":
          description: Saved session with fresh compute result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZakatSessionCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/zakat/sessions/{id}:
    get:
      tags: [Consumer]
      summary: Load a saved zakat session
      operationId: getZakatSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session inputs and result snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZakatSessionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Consumer]
      summary: Delete a saved zakat session
      operationId: deleteZakatSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    BadRequest:
      description: Validation or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Unauthorized:
      description: Missing or invalid credentials / token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    ServiceUnavailable:
      description: Upstream unavailable and no acceptable cache
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    WatchlistAddRequest:
      type: object
      description: Provide instrumentId or symbol (EGX ticker); symbol creates/links a DB instrument when missing.
      properties:
        instrumentId:
          type: string
          format: uuid
        symbol:
          type: string
        displayNameEn:
          type: string
        displayNameAr:
          type: string
    WatchlistReorderRequest:
      type: object
      required: [orderedItemIds]
      properties:
        orderedItemIds:
          type: array
          items:
            type: string
            format: uuid
    WatchlistItem:
      type: object
      required:
        [
          id,
          instrumentId,
          position,
          code,
          displayNameEn,
          kind,
          isConsumerVisible,
        ]
      properties:
        id:
          type: string
          format: uuid
        instrumentId:
          type: string
          format: uuid
        position:
          type: integer
        code:
          type: string
        displayNameEn:
          type: string
        kind:
          type: string
        isConsumerVisible:
          type: boolean
    WatchlistListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WatchlistItem"
...
    ZakatNisab:
      type: object
      required:
        - disclaimer
        - methodologyNote
        - nisabGoldGrams
        - nisabGoldKarat
        - gold21GramPriceEgp
        - nisabEgp
        - asOf
        - isStale
        - rates
      properties:
        disclaimer:
          type: string
        methodologyNote:
          type: string
        nisabGoldGrams:
          type: number
        nisabGoldKarat:
          type: integer
        gold21GramPriceEgp:
          type: number
        nisabEgp:
          type: number
        asOf:
          type: string
          format: date-time
        isStale:
          type: boolean
        rates:
          type: object
          required: [hijri, gregorian]
          properties:
            hijri:
              type: number
            gregorian:
              type: number
    ZakatComputeRequest:
      type: object
      required: [yearType]
      properties:
        yearType:
          type: string
          enum: [hijri, gregorian]
        nisabAttainmentDate:
          type: string
          format: date
        cashEgp:
          type: number
          minimum: 0
        equitiesEgp:
          type: number
          minimum: 0
        equitiesPurpose:
          type: string
          enum: [trading, long_term_investment]
        gold:
          type: object
          properties:
            mode:
              type: string
              enum: [egp, grams]
            amount:
              type: number
              minimum: 0
            karat:
              type: integer
              enum: [18, 21, 24]
            purpose:
              type: string
              enum: [investment, personal_jewelry]
        silver:
          type: object
          properties:
            mode:
              type: string
              enum: [egp, grams]
            amount:
              type: number
              minimum: 0
        bankCertificatesEgp:
          type: number
          minimum: 0
        realEstateEgp:
          type: number
          minimum: 0
        commercialAssetsEgp:
          type: number
          minimum: 0
        receivablesEgp:
          type: number
          minimum: 0
        debtsOwedEgp:
          type: number
          minimum: 0
    ZakatComputeResponse:
      allOf:
        - $ref: "#/components/schemas/ZakatNisab"
        - type: object
          required:
            - yearType
            - rate
            - breakdown
            - aboveNisab
            - zakatDueEgp
            - hawlReminder
          properties:
            yearType:
              type: string
              enum: [hijri, gregorian]
            rate:
              type: number
            nisabAttainmentDate:
              type: string
              format: date
              nullable: true
            breakdown:
              type: object
            aboveNisab:
              type: boolean
            zakatDueEgp:
              type: number
            hawlReminder:
              type: string
            hawl:
              type: object
              nullable: true
            purposeNotes:
              type: array
              items:
                type: object
...
    ZakatSessionSummary:
      type: object
      required: [id, yearType, zakatDueEgp, aboveNisab, createdAt]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          nullable: true
        yearType:
          type: string
        nisabAttainmentDate:
          type: string
          format: date
          nullable: true
        zakatDueEgp:
          type: number
        aboveNisab:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ZakatSessionCreate:
      type: object
      required: [inputs]
      properties:
        label:
          type: string
          maxLength: 120
          nullable: true
        inputs:
          $ref: "#/components/schemas/ZakatComputeRequest"
    ZakatSessionCreateResponse:
      type: object
      required: [session, result]
      properties:
        session:
          $ref: "#/components/schemas/ZakatSessionSummary"
        result:
          $ref: "#/components/schemas/ZakatComputeResponse"
    ZakatSessionDetail:
      type: object
      required: [id, inputs, result, createdAt]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          nullable: true
        inputs:
          type: object
        result:
          type: object
        createdAt:
          type: string
          format: date-time
    ZakatPrefill:
      type: object
      required: [disclaimer, equities, equitiesTotalEgp]
      properties:
        disclaimer:
          type: string
        equitiesTotalEgp:
          type: number
        equities:
          type: array
          items:
            type: object
            required: [instrumentId, code, displayNameEn, marketValueEgp]
            properties:
              instrumentId:
                type: string
                format: uuid
              code:
                type: string
              displayNameEn:
                type: string
              marketValueEgp:
                type: number